RaterhubAutomator/requirements.txt
selenium
webdriver_manager
Pillow
PySimpleGUI
pywin32
google-cloud-aiplatform
google-generativeai
